Explicitly cast to NSInteger

+++ Code/Network/RKResponse.m
@@ -209,7 +209,7 @@ return __VA_ARGS__;                                                             
     [_body appendData:data];
     [_request invalidateTimeoutTimer];
     if ([[_request delegate] respondsToSelector:@selector(request:didReceiveData:totalBytesReceived:totalBytesExpectedToReceive:)]) {
-        [[_request delegate] request:_request didReceiveData:[data length] totalBytesReceived:[_body length] totalBytesExpectedToReceive:_httpURLResponse.expectedContentLength];
+        [[_request delegate] request:_request didReceiveData:[data length] totalBytesReceived:[_body length] totalBytesExpectedToReceive:(NSInteger)_httpURLResponse.expectedContentLength];
     }
 }
